rename file name in page dir
